Fixed bug: winding of contours sometimes filled path inverted
Fix is possibly a bit ad-hoc, but should be robust: check if some trail point
taken from a decision boundary path is "inside" the path under the "evenodd"
fill rule. If not, add a rect that encompasses the whole canvas to the path,
effectively alternating which areas are filled and which aren't.